requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0 
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
openpyxl
pandas
pyautogui